+++ CHANGELOG.md
@@ -4,14 +4,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `className` prop propagation to `SingleEntryTable` and subcomponents:
+## [2.124.0] - 2021-03-12
+
+Features:
+
+- Add `className` prop propagation to `SingleEntryTable` and subcomponents:
   - `SingleEntryTable.Col`
   - `SingleEntryTable.HeaderCol`
-- Feature: Add `className` prop propagation to `DoubleEntryTable` and subcomponents:
+- Add `className` prop propagation to `DoubleEntryTable` and subcomponents:
   - `DoubleEntryTable.Col`
   - `DoubleEntryTable.HeaderCol`
   - `DoubleEntryTable.TitleCol`
-
 
 ## [2.123.3] - 2021-03-10